Fixing the problem with module import from pytest
