+++ pyproject.toml
@@ -14,6 +14,11 @@ readme = "README.md"
 requires-python = ">=3.10"
 dynamic = ["dependencies", "optional-dependencies"]
 
+[tool.pytest.ini_options]
+pythonpath = [
+  "mlopstinystories"
+]
+
 [tool.setuptools]
 packages = ["mlopstinystories"]
